# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Prueba Tecnica Instaleap",
    "version": "1.0.0",
    "description": "Documentación de API para manejo de usuarios"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    },
    {
      "url": "https://prueba-tecnica-fos4.onrender.com",
      "description": "Produccion"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "67c33824326cb76fdfb91434"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "secretpassword"
          },
          "roles": {
            "type": "string",
            "example": [
              "admin",
              "user",
              "editor"
            ]
          },
          "edad": {
            "type": "number",
            "example": 20
          }
        }
      },
      "Tarea": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "67c33824326cb76fdfb91434"
          },
          "titulo": {
            "type": "string",
            "example": "Tarea # 1"
          },
          "descripcion": {
            "type": "string",
            "example": "Esta es la tarea # 1"
          },
          "fecha_vencimiento": {
            "format": "date",
            "type": "string",
            "example": "2025-03-05"
          },
          "estado": {
            "type": "string",
            "example": [
              "Pendiente",
              "Completada",
              "Cancelada"
            ],
            "items": {
              "type": "string"
            }
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario",
            "example": "67c33824326cb76fdfb91434"
          }
        }
      }
    }
  },
  "paths": {
    "/registrar": {
      "post": {
        "summary": "Permite el registro de un usuario",
        "tags": [
          "Usuarios"
        ],
        "responses": {
          "201": {
            "description": "Usuario regisrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "500": {
            "description": "Error: Mensaje del sistema de validacion de entredas o del servicio."
          }
        }
      }
    },
    "/profile": {
      "get": {
        "summary": "Permite el acceso a los datos de usuario.",
        "tags": [
          "Usuarios"
        ],
        "responses": {
          "200": {
            "description": "[JSON] Usuario encontrado - Datos de usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Error: Usuario no encontrado"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Permite el acceso al listado de usuarios con autorizacion [admin].",
        "tags": [
          "Usuarios"
        ],
        "responses": {
          "200": {
            "description": "[JSON] arreglo de usuarios registrados en el sistema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Error: Este error no tiene sentido, sino quien haria la consulta?... jejejejeje"
          }
        }
      }
    },
    "/{id}": {
      "delete": {
        "summary": "Permite al usuario eliminar una tarea (mientras los datos sean validados, autoria de la tarea)",
        "tags": [
          "Tareas"
        ],
        "responses": {
          "200": {
            "description": "La tarea con el identificador ################## fue eliminada con exito.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tarea"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error: Mensaje del sistema de validacion de entredas o del servicio."
          }
        }
      },
      "put": {
        "summary": "Permite al usuario editar una tarea (mientras los datos sean validados solo requiere los campos que se vayan a modificar)",
        "tags": [
          "Tareas"
        ],
        "responses": {
          "202": {
            "description": "[JSON] Datos de la tarea editada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tarea"
                }
              }
            }
          },
          "500": {
            "description": "Error: Mensaje del sistema de validacion de entredas o del servicio."
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Permite a los usuarios identificarse con el sistema y recibir un token JWT para identificarse en proximas solicitudes, limita el numero (5) de intentos fallidos para loggearse.",
        "tags": [
          "Usuarios"
        ],
        "responses": {
          "200": {
            "description": "[JSON] con los datos de usuario y token generado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Email y constraseña son requeridos"
          },
          "401": {
            "description": "Datos de usuario invalidos"
          },
          "404": {
            "description": "Usuario no encontrado."
          }
        }
      }
    },
    "/tareas": {
      "get": {
        "summary": "[JSON] Permite al usuario listar las tareas registradas por el usuario.",
        "tags": [
          "Tareas"
        ],
        "responses": {
          "201": {
            "description": "Listado de tareas registradas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tarea"
                }
              }
            }
          },
          "500": {
            "description": "Error: Mensaje del sistema de validacion de entredas o del servicio."
          }
        }
      }
    },
    "/addTarea": {
      "post": {
        "summary": "Permite al usuario registrar una tarea",
        "tags": [
          "Tareas"
        ],
        "responses": {
          "201": {
            "description": "[JSON] Datos de la tarea registrada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tarea"
                }
              }
            }
          },
          "500": {
            "description": "Error: Mensaje del sistema de validacion de entredas o del servicio."
          }
        }
      }
    }
  },
  "tags": []
}